requirements.txt
flake8
pillow
numpy
matplotlib
opencv-python #to cv2